fix: rename package for (internal) ExoPlayerAdapter classes
