+++ ExoPlayerAdapter/src/exo-analytics-listener-2_13-2_16/java/com/mux/stats/sdk/muxstats/internal/ExoAnalyticsListener.kt
@@ -11,6 +11,11 @@ import com.google.android.exoplayer2.source.MediaLoadData
 import com.google.android.exoplayer2.source.TrackGroupArray
 import com.google.android.exoplayer2.trackselection.TrackSelectionArray
 import com.mux.stats.sdk.muxstats.MuxStateCollector
+import com.mux.stats.sdk.muxstats.exoplayeradapter.internal.BandwidthMetricDispatcher
+import com.mux.stats.sdk.muxstats.exoplayeradapter.internal.MuxMediaHasVideoTrack
+import com.mux.stats.sdk.muxstats.exoplayeradapter.internal.handleExoPlaybackState
+import com.mux.stats.sdk.muxstats.exoplayeradapter.internal.handlePositionDiscontinuityBefore218
+import com.mux.stats.sdk.muxstats.exoplayeradapter.internal.weak
 import java.io.IOException
 
 /**
