+++ ExoPlayerAdapter/src/exo-analytics-upto-2_16/java/com/mux/stats/sdk/muxstats/AnalyticsListenerMetrics.kt
@@ -3,10 +3,11 @@ package com.mux.stats.sdk.muxstats
 import com.google.android.exoplayer2.SimpleExoPlayer
 import com.google.android.exoplayer2.analytics.AnalyticsListener
 import com.mux.stats.sdk.core.util.MuxLogger
+import com.mux.stats.sdk.muxstats.exoplayeradapter.MuxPlayerAdapter
 import com.mux.stats.sdk.muxstats.internal.exoAnalyticsListener
-import com.mux.stats.sdk.muxstats.internal.logTag
-import com.mux.stats.sdk.muxstats.internal.watchContentPosition
-import com.mux.stats.sdk.muxstats.internal.weak
+import com.mux.stats.sdk.muxstats.exoplayeradapter.internal.logTag
+import com.mux.stats.sdk.muxstats.exoplayeradapter.internal.watchContentPosition
+import com.mux.stats.sdk.muxstats.exoplayeradapter.internal.weak
 
 private class AnalyticsListenerBindingUpTo16 : MuxPlayerAdapter.PlayerBinding<SimpleExoPlayer> {
