+++ ExoPlayerAdapter/src/exo-error-just-2_14/java/com/mux/stats/sdk/muxstats/ExoErrorMetricsUpTo214.kt
@@ -4,9 +4,10 @@ import com.google.android.exoplayer2.ExoPlaybackException
 import com.google.android.exoplayer2.ExoPlayer
 import com.google.android.exoplayer2.Player
 import com.mux.stats.sdk.core.util.MuxLogger
-import com.mux.stats.sdk.muxstats.internal.handleExoPlaybackException
-import com.mux.stats.sdk.muxstats.internal.logTag
-import com.mux.stats.sdk.muxstats.internal.weak
+import com.mux.stats.sdk.muxstats.exoplayeradapter.MuxPlayerAdapter
+import com.mux.stats.sdk.muxstats.exoplayeradapter.internal.handleExoPlaybackException
+import com.mux.stats.sdk.muxstats.exoplayeradapter.internal.logTag
+import com.mux.stats.sdk.muxstats.exoplayeradapter.internal.weak
 
 /**
  * Player binding for  exoplayer android metrics
