+++ ExoPlayerAdapter/src/test/java/com/mux/exoplayeradapter/double/FakePlayerBinding.kt
@@ -1,9 +1,9 @@
 package com.mux.exoplayeradapter.double
 
 import com.mux.exoplayeradapter.log
-import com.mux.stats.sdk.muxstats.MuxPlayerAdapter
+import com.mux.stats.sdk.muxstats.exoplayeradapter.MuxPlayerAdapter
 import com.mux.stats.sdk.muxstats.MuxStateCollector
-import com.mux.stats.sdk.muxstats.internal.logTag
+import com.mux.stats.sdk.muxstats.exoplayeradapter.internal.logTag
 
 class FakePlayerBinding<Player>(val name: String) : MuxPlayerAdapter.PlayerBinding<Player> {
   override fun bindPlayer(player: Player, collector: MuxStateCollector) {
